requirements.txt
Flask>=2.3.0
pymongo>=4.0.0
pytesseract>=0.3.0
Pillow>=10.0.0
pdf2image>=1.16.0
python-docx>=0.8.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
gunicorn>=21.0.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
werkzeug>=2.3.0
marshmallow>=3.20.0
flasgger>=0.9.0
opencv-python>=4.8.0